Add files via upload
This code is a Python script that uses the Tweepy library to search for 10 tweets containing the keyword "covid" on Twitter. It connects to the Twitter API using authentication credentials stored in a separate file called "config.py". The script then prints out the ID and language of each tweet that matches the search criteria.